fix: App과 Route를 분리하고 lazy loading을 적용함
